xfix/requirements.txt
# XFix dependencies
httpx>=0.27.0
python-dotenv>=1.0.0
ollama>=0.4.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
